fix(Button): change minWidth (#709)
Modifying min button width size and fixes snapshots based on this
modification, requested by Vivo.

Figma link:
https://www.figma.com/file/koROdh3HpEPG2O8jG52Emh/%F0%9F%94%B8-Buttons-Component-Specs?node-id=0-1&t=eJRUzlQqlKF9xRBL-0

Important note: Changing this min-width value, as it is proposed by
default behavior, changes several already committed snapshots. Initially
we noticed the 4 around the button in a direct way, in which it was
indicated to change. But opening the PR and retesting completely
locally, another 70 snapshots showed errors, due to the button layout
change. I would like you to know the most appropriate thing to do in
this case. Would it change all the other snapshots and commit to this
PR? Is there any indication for this case?

---------

Co-authored-by: Abel Toledano <atoledano@tuenti.com>