fix(TextField, Select): improve styles (#777)
- [x] perfect alignment in select input with text field in chrome
- [x] perfect alignment in select input with text field in safari 
- [x] Changed input password text to use Lucida Grande instead of Arial
in MacOS
- [x] Changed icons in CVV tooltips to use Mística colors and use text
presets
- [x] Changed icon in credit card field to use mistica icons. Removed
the old one.
- [x] Changed search field icon to use mistica icons. Removed the old
one.

⚠️ I changed the inherited fontSize and lineHeight in input field, now,
input fields don't respond to font scale. We should improve this
behavior in the future to adapt input box to the font size. I did it to
make parity between select and input fields. Also because the behavior
of text with scale in input fields is not working as best

alignment test → https://tinyurl.com/23u5s55u

---
Chrome

![image](https://github.com/Telefonica/mistica-web/assets/6722153/8eb27d10-7f94-42be-968a-8666c82083b8)

Safari

![image](https://github.com/Telefonica/mistica-web/assets/6722153/65b67d0a-4a4e-4436-8d66-28a91706765a)