feat(Header): small version, allow removing vertical padding (#822)
Jira with specs: https://jira.tid.es/browse/WEB-1469

- Added a `small` prop in `Header` to reduce the title/description font
size
- Added a `noPaddingY` prop to `HeaderLayout` to remove the vertical
padding (required by Novum CMS layout, which establishes its own
paddings)
- Added support for responsive space in `Stack` and `Box` components:
```ts
<Stack space={{mobile: 8, tablet: 16, desktop: 24}}>

<Box paddingY={{mobile: 16, desktop: 24}}>
```
When using the responsive notation (object), setting both `mobile` and
`desktop` is mandatory. `tablet` is optional, and will use `mobile`
value by default when not specified.
This will allow us to reduce the use of `useScreenSize()` hook, which is
problematic with SSR.
In this PR I've applied this to header components, but we should do the
same with the rest of components in Mística too. It's still pending to
apply the same approach in `Inline` component