feat(utilities): Remove lodash dependency and replace it with custom functions (#856)
This will help us to make mistica a fully ES Module library. In my
tests, when changing the package type to `module`, the build was
complaining about `lodash` (which is a cjs library).

I tried to replace it with `lodash-es` but for some reason that I was
unable to understand, that produced problems with the SSR acceptance
tests, so at the end we decided to fully remove lodash from mistica
components (and we saved 2.2 kB)

---------

Co-authored-by: Pedro Ladaria <pedro.jose.ladaria.linden@telefonica.com>
Co-authored-by: Abel Toledano <atoledano@tuenti.com>