feat(Row): improve VoiceOver compatibility (#1079)
Fixes:
WEB-1724 
WEB-1740

The main issue that was causing the bugs is that Safari reads links with
multiple lines like different links. The workaround for this is:

- Add an aria-label to the `<a>` element
- Add an aria-hidden to the `<a>` children
- The aria-label of the `<a>` should be computed with the content of the
row, in this case we are concatenating `[title, headline, subtitle,
description, extra]` props to create the aria-label. The problem is that
some of these props aren't simple `string`s but any `ReactNode`, so we
need to extract the inner text of the html elements on runtime to
compute the `aria-label` for the `<a>`

Other changes in the PR:
- Fix some cases where the aria-label was not being set properly in rows
with controls (checkbox/radio/switch)
- Change the dom order of `headline` to make screen readers read it
after `title`

Still pending:
- Review badge a11y
- Review how should screen reader read `details` part. When the row is a
link/button it is read as part of the row as expected. But when we use
some control (switch/checkbox/radio) the details part is not read
because it's outside the div referenced by `aria-labeledby`. Including
`details` in the same div used for `headline`, `title`, `description`,
etc would require a big refactor of this component layout