feat(Row): allow aria-label in informative rows (#1269)
https://jira.tid.es/browse/WEB-2061
By default, the browser will read text content inside rows one by one
(title, description, etc).
With the changes in this PR, if an aria-label is provided to the
informative row, the whole row will be read as a single text (the one
provided in the aria-label prop)

This PR also fixes https://jira.tid.es/browse/WEB-1885 which is a
basically revert of https://github.com/Telefonica/mistica-web/pull/942.
As part of this PR we allow passing custom styles (via `className` prop)
to `ScreenReaderOnly` to workaround possible style issues found (like
the one described originally in
https://github.com/Telefonica/mistica-web/pull/942).