fix: :bug: dynamically handle scroll-padding-bottom for fixed footer
This commit implements the necessary changes to dynamically handle scroll-padding-bottom on the html element. This ensures that the fixed footer does not block the focus of overflow elements when visible.

Added a <style> tag to dynamically set  scroll-padding-bottom on the html element based on the footer height.

Related:
This commit is related to [O2DE-6995](https://jira.tid.es/browse/O2DE-6995).