fix(FixedFooterLayout): obscured focus (#1334)
fix(FixedFooterLayout): fixed footer layout obscured focus
* This pull request includes a small change to the `FixedFooterLayout`
component in `src/fixed-footer-layout.tsx`. The change ensures that when
the footer is visible, a style is applied to set the
`scroll-padding-bottom` to the height of the footer.

*
[`src/fixed-footer-layout.tsx`](diffhunk://#diff-a20f24af3140080300870a68a5accb89feee523f1c59592a9e57d0f2bf4d02baR166):
Added a style element to set `scroll-padding-bottom` when the footer is
visible.
This commit is related to
[[O2DE-6995](https://jira.tid.es/browse/O2DE-6995)].